Update java
Hi, please use a stack instead of a queue. We will traverse the list from last to keep track of the trees with visible height. Whenever we encounter trees smaller than the current height we will pop them from the stack as they won't be visible when we look at them from the front. Do let me know if I can help you understand it better.